refactor: Move dsl into queries directory
